Add better status reporting using CATMAID Messages and WebSockets
First, the circuit map widget uses now the general widget content
messages to display some status information what it currently does (grey
text in background of widget content pane). This is longer visible than
the growl messages, which can be easy to miss at times.

If the initial request doesn't lead to the queuing of an import task,
proper HTTP errors are now returned and the front-end can handle them
bitter. A new error type hs been introduced so that the front-end can
distinguish between circuit map processing errors (e.g. no segment found
for a location) and other errors. Circuit map errors will be display by
the front-end only as warning (and hence a growl message) and the
content background message.

The back-end will now also emit WebSockets message to all front-ends of
the user who queued the request. This happens both for errors and
successes and the front-end can react to events more quickly. In the
case of succss of the import of synapses for an existing skeleton, a
skeleton update event is now emitted in the front-end, allowing other
widgets to reload the potentially changed skeleton (and its
connectivity). In the case of a successful import at a location (i.e. no
active skeleton), the views are now reloaded automatically, so that the
new skeleton becomes visible immediately. In case of error, the error
handling as described above is performed.

Additionally, for each task result a CATMAID Message is created so that
setups without WebSockets/ASGI can get update information too. They
appear in CATMAID's message menu an are marked as read by clicking on
them.