Import: Use connector node to link to skeleton
Before, the presynaptic site was used to find the closest skeleton node to link to. When afterwards the original presynaptic site was replaced with a connectore node, this resulted in different skeleton nodes being linked to the same connector node. Fix this bug by using the connector node in the first place (not tested!).