Fix treenode ID referencing variable name
The previous name of "skid" alludes to "skeleton ID", a term commonly
used in CATMAID to refer to whole skeletons. The code however didn't
refer to actual skeletons but to skeleton nodes. To stay consistent with
the rest of CATMAID, "treenode_id" is used instead. This helps the
readability of the code.