Import partner fragments in same transaction
So far we imported partner fragments in separate Celery tasks, and hence
separate transactions. This can be confusing with respect to when the
containing task finishes. Now that we have task tracking tables, it is
easier to keep track of progress. So even if an individual task is now
running longer rather than spawning new tasks, the users is updated
about progress properly.